Moved Linear and NonLinear module definitions to header files
